LayoutTests/http/tests/media/video-play-stall-expected.txt


Test that a stalled event is sent when media loading stalls.

EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
RUN(video.play())
EVENT(stalled)
EXPECTED (video.currentTime != '0') OK
EXPECTED (video.playbackRate === '1') OK
EXPECTED (video.paused === 'false') OK
END OF TEST
